apps/api/requirements.txt
fastapi==0.115.5
uvicorn==0.30.6
python-multipart==0.0.9
pypdf==4.3.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg[binary]==3.2.3
python-dotenv>=1.2.1
azure-storage-blob==12.24.1
azure-identity==1.19.0
